Do not open links in new windows
The change to open links in new windows was made claiming "security
issues", however no actual security issues were presented in the
original issue[1][2]. There does exist a vulnerability with
`window.opener`[3], however that issue only exists when using an `<a>`
with a non-empty `target` attribute or calling `window.open` [4].

Forcing the browser to open links in a new tab is bad user experience,
as it breaks the behavior of the back button and needlessly creates
additional tabs/windows. The user can usually easily decide to open a
link in a new tab themselves, but there is typically no straightforward
way for a user to force a link to open in the *same* tab.

Further, many new external links without a `target="_blank"` have been
added since the original change (over 90 using a rough count), causing
inconsistency in how external links behave, and showing that it, as
implemented, is not a  sustainable pattern. Thus it is simpler to use
the default behavior of `<a>` (no `target`), which remains as safe as
specifying a `target` with `rel="noopener"`.

[1] https://github.com/rust-lang/www.rust-lang.org/issues/246
[2] https://github.com/rust-lang/www.rust-lang.org/pull/263
[3] https://mathiasbynens.github.io/rel-noopener/
[4] https://developer.mozilla.org/en-US/docs/Web/API/Window/opener#Notes